% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ActionButton.R
\name{ActionButton}
\alias{ActionButton}
\title{ActionButton}
\usage{
ActionButton(app, inputId, label, icon = NULL, width = NULL, ...)
}
\arguments{
\item{app}{ShinyApp object}

\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{The contents of the button or link--usually a text label, but
you could also use any other HTML, like an image.}

\item{icon}{An optional \code{\link{icon}} to appear on the button.}

\item{width}{The width of the input, e.g. \code{'400px'}, or \code{'100\%'};
see \code{\link{validateCssUnit}}.}

\item{...}{Named attributes to be applied to the button or link.}
}
\value{
an updated app
}
\description{
Creates an action button element.
}
\examples{
\dontrun{ShinyApp() \%>\% ActionButton('button', 'Button label') }
}
