% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add7Nwatermark.R
\name{add7NWatermark}
\alias{add7NWatermark}
\title{Add 7N watermark to graph}
\usage{
add7NWatermark(file.name, file.width, file.height,
 watermark.path, x.coordinate, y.coordinate, width, plot.unit)
}
\arguments{
\item{file.name}{name of the file}

\item{file.width}{name of the graph PNG file}

\item{file.height}{directory to which the graph should be saved to}

\item{watermark.path}{directory to the watermark logo}

\item{x.coordinate}{x coordinates of the position of watermark}

\item{y.coordinate}{y coordinates of the position of watermark}

\item{width}{width of the watermark}

\item{plot.unit}{unit used in function}
}
\value{
png with added watermark
}
\description{
\code{add7NWatermark} adds 7N watermark to given graph
}
\examples{
test = system.file("extdata", "test.png", package = "ShinyApp")
logo = system.file("extdata", "logo.png", package = "ShinyApp")
add7NWatermark(test, 1200, 800, logo, 0.84, 0.17, 0.02, 'npc')
}
