% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadData.R
\docType{data}
\name{baseUrl}
\alias{baseUrl}
\title{Base url to gather data from}
\format{An object of class \code{character} of length 1.}
\usage{
baseUrl
}
\arguments{
\item{baseUrl}{contains the URL for API Call}
}
\description{
Base url to gather data from
}
\keyword{datasets}
