% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckboxInput.R
\name{CheckboxInput}
\alias{CheckboxInput}
\title{CheckboxInput}
\usage{
CheckboxInput(app, inputId, label, value = FALSE, width = NULL)
}
\arguments{
\item{app}{ShinyApp object}

\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{Display label for the control, or \code{NULL} for no label.}

\item{value}{Initial value (\code{TRUE} or \code{FALSE}).}

\item{width}{The width of the input, e.g. \code{'400px'}, or \code{'100\%'};
see \code{\link{validateCssUnit}}.}
}
\value{
an updated app
}
\description{
Creates a checkbox element.
}
\examples{
\dontrun{ShinyApp() \%>\% CheckboxInput('rug', 'Show observations') }
}
