% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dashboard.R
\name{DashboardHeader}
\alias{DashboardHeader}
\title{DashboardHeader}
\usage{
DashboardHeader(app, title = NULL, titleWidth = NULL,
  disable = FALSE)
}
\arguments{
\item{app}{a ShinyApp object}

\item{title}{An optional title to show in the header bar.. By default, this
will also be used as the title shown in the browser's title bar. If you
want that to be different from the text in the dashboard header bar, set
the \code{title} in \code{\link{dashboardPage}}.}

\item{titleWidth}{The width of the title area. This must either be a number
which specifies the width in pixels, or a string that specifies the width
in CSS units.}

\item{disable}{If \code{TRUE}, don't display the header bar.}
}
\value{
an updated ShinyApp object
}
\description{
DashboardHeader
}
\examples{
\dontrun{ShinyApp \%>\% DashboardHeader('MM*Stat')}
}
