% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dashboard.R
\name{DashboardSidebar}
\alias{DashboardSidebar}
\title{DashboardSidebar}
\usage{
DashboardSidebar(app, disable = FALSE, width = NULL,
  collapsed = FALSE)
}
\arguments{
\item{app}{a ShinyApp object}

\item{disable}{If \code{TRUE}, the sidebar will be disabled.}

\item{width}{The width of the sidebar. This must either be a number which
specifies the width in pixels, or a string that specifies the width in CSS
units.}

\item{collapsed}{If \code{TRUE}, the sidebar will be collapsed on app startup.}
}
\value{
an updated ShinyApp object
}
\description{
DashboardSidebar
}
\examples{
#' \dontrun{ShinyApp \%>\% DashboardSidebar(collapsed = TRUE)}
}
