% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadData.R
\name{datasuccess}
\alias{datasuccess}
\title{Connects to the API and returns data}
\usage{
datasuccess()
}
\value{
jsndata - A json object
}
\description{
Connects to the API and returns data
}
\examples{
 
datasuccess()
}
