% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTimeSeriesToPlot.R
\name{getTimeSeriesToPlot}
\alias{getTimeSeriesToPlot}
\title{Convert time series to data frame used in plotting graphs}
\usage{
getTimeSeriesToPlot(time.series)
}
\arguments{
\item{time.series}{a time series with data}
}
\value{
data frame with transformed time series
}
\description{
\code{getTimeSeriesToPlot} converts time series to data frame used in plotting graphs
}
\examples{
getTimeSeriesToPlot(datasets::AirPassengers)

}
