% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Global.R
\name{Global}
\alias{Global}
\title{Global}
\usage{
Global(app, expr, file = NULL)
}
\arguments{
\item{app}{ShinyApp object}

\item{expr}{expression: code to add to the app}

\item{file}{chracter: file name of a file which contains the code}
}
\value{
an updated ShinyApp object
}
\description{
Allows to enter global code into the app. Please note that in shiny server not all commands allowed. 
For example using \code{install.packages} will
}
\examples{
\dontrun{
  ShinyApp() \%>\% 
    Global({
      library("rio")
      x <- import("mydata")
    })
}
}
