% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui.R
\name{launchApp}
\alias{launchApp}
\title{shiny app
This app allows you to visualize the functions implemented in the algorithms package.}
\usage{
launchApp()
}
\description{
shiny app
This app allows you to visualize the functions implemented in the algorithms package.
}
