% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotConditionalGraph.R
\name{plotConditionalGraph}
\alias{plotConditionalGraph}
\title{Plots conditional graph}
\usage{
plotConditionalGraph(time.series, months, condition)
}
\arguments{
\item{time.series}{a time series with data}

\item{months}{number of months before date to compare}

\item{condition}{percent of growth/decrease comparing to months before}
}
\value{
points to add on graph plot
}
\description{
\code{plotConditionalGraph} plots conditional graph
}
\examples{
plotConditionalGraph(datasets::AirPassengers, 12, 0.15)
}
