% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGraph.R
\name{plotGraph}
\alias{plotGraph}
\title{Plot graph from data}
\usage{
plotGraph(time.series, name, directory)
}
\arguments{
\item{time.series}{a time series with data}

\item{name}{name of the graph PNG file}

\item{directory}{directory to which the graph should be saved to}
}
\value{
png with graph
}
\description{
\code{plotGraph} plots graph of given time series
}
\examples{
plotGraph(datasets::AirPassengers, 'timeSeries', 'files')
}
