% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveToCsv.R
\name{saveToCsv}
\alias{saveToCsv}
\title{Saves data to csv}
\usage{
saveToCsv(time.series, name, directory)
}
\arguments{
\item{time.series}{a time series with data}

\item{name}{name of the file}

\item{directory}{directory to save the file}
}
\value{
csv file
}
\description{
\code{saveToCsv} saves time series data to CSV
}
\examples{
saveToCsv(datasets::AirPassengers, name = 'fanpage2', directory = 'files')
}
