% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{server}
\alias{server}
\title{Shiny Server for the Interactive Data Analysis Application}
\usage{
server(input, output)
}
\value{
A reactive Shiny server function that filters and displays data based on user input.
}
\description{
This is the server component of the Shiny app. It handles data downloading, filtering, and displaying the data in a table format.
}
