\name{Shinyapp-package}
\alias{Shinyapp-package}
\alias{Shinyapp}
\docType{package}
\title{
\packageTitle{Shinyapp}
}
\description{
\packageDescription{Shinyapp}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{Shinyapp}

Maintainer: \packageMaintainer{Shinyapp}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = Shinyapp)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{Shinyapp}}
%\section{Documentation Index}{\packageIndices{Shinyapp}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
