% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyApp.R
\name{ShinyApp}
\alias{ShinyApp}
\alias{ShinyApp.default}
\title{ShinyApp}
\usage{
ShinyApp(app, ...)

\method{ShinyApp}{default}(app, ...)
}
\arguments{
\item{app}{character: name of app template file (default: \code{inst/app/app.tmpl} in the \code{shinyExample} library)}

\item{...}{unused}
}
\value{
an empty ShinyApp object
}
\description{
Creates a shiny app using \href{https://rstudio.github.io/shinydashboard/}{shinydashboard}.
}
\examples{
ShinyApp()
}
