% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextOutput.R
\name{TextOutput}
\alias{TextOutput}
\title{TextOutput}
\usage{
TextOutput(app, outputId, container = if (inline) span else div,
  inline = FALSE, file, text, expr)
}
\arguments{
\item{app}{a ShinyApp object}

\item{outputId}{output variable to read the value from}

\item{container}{a function to generate an HTML element to contain the text}

\item{inline}{use an inline (\code{span()}) or block container (\code{div()})
for the output}

\item{file}{character: name of file with R code to execute}

\item{text}{character: text with R code to execute}

\item{expr}{expression: R code to execute}
}
\value{
an updated ShinyApp object
}
\description{
Render a reactive output variable as text within an application page. 
Either \code{file} or \code{text} must be set.
}
\examples{
\dontrun{ShinyApp() \%>\% TextOutput('plot', text='"Hello, World!"' }
}
