% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turnTimeSeriesToDataFrame.R
\name{turnTimeSeriesToDataFrame}
\alias{turnTimeSeriesToDataFrame}
\title{Convert time series to data frame}
\usage{
turnTimeSeriesToDataFrame(time.series)
}
\arguments{
\item{time.series}{a time series with data}
}
\value{
data frame with transformed time series
}
\description{
\code{turnTimeSeriesToDataFrame} converts time series to data frame
}
\examples{
turnTimeSeriesToDataFrame(datasets::AirPassengers)

}
