% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\docType{data}
\name{ui}
\alias{ui}
\title{Shiny UI for the Interactive Data Analysis Application}
\format{
An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 4.
}
\usage{
ui
}
\value{
Returns the user interface for the Shiny application.
}
\description{
This is the UI component of the Shiny app. It contains two inputs: one for selecting the year and another for filtering values greater than a selected threshold. The filtered data is displayed in a table format.
}
\keyword{datasets}
